+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Workaround for mods with incorrect drop or recipe `ColumnNum` values
+- Workaround for mods with incorrect drop/recipe numbered names (e.g. `ItemNum001`)
 - Queue errors for end of batch install
 - More error details, especially to identify mods in batch operations