+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.2]
+
 ### Changed
 
 - Workaround for mods with incorrect drop or recipe `ColumnNum` values
 - Workaround for mods with incorrect drop/recipe numbered names (e.g. `ItemNum001`)
 - Queue errors for end of batch install
 - More error details, especially to identify mods in batch operations
+- Clearer "no base version" error
+
+### Fixed
+
+- Hacky fix for low Switch RSTB values
 
 ## [0.8.1]